third_party/blink/web_tests/platform/generic/external/wpt/css/css-contain/container-queries/fragmented-container-001-expected.txt

This is a testharness.js-based test.
FAIL Children of fragmented inline-size container should match inline-size of first fragment assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
